Moved file stuff into the box package
