+++ qabelbox/src/main/java/de/qabel/qabelbox/box/adapters/FileAdapter.kt
@@ -1,4 +1,4 @@
-package de.qabel.qabelbox.ui.adapters
+package de.qabel.qabelbox.box.adapters
 
 import android.support.v7.widget.RecyclerView
 import android.view.LayoutInflater
