+++ qabelbox/src/main/java/de/qabel/qabelbox/dagger/components/MainActivityComponent.java
@@ -2,12 +2,13 @@ package de.qabel.qabelbox.dagger.components;
 
 import dagger.Subcomponent;
 import de.qabel.qabelbox.activities.MainActivity;
+import de.qabel.qabelbox.box.dagger.modules.FileBrowserComponent;
 import de.qabel.qabelbox.contacts.dagger.ContactDetailsComponent;
 import de.qabel.qabelbox.contacts.dagger.ContactDetailsModule;
 import de.qabel.qabelbox.contacts.dagger.ContactsComponent;
 import de.qabel.qabelbox.contacts.dagger.ContactsModule;
 import de.qabel.qabelbox.dagger.modules.ChatModule;
-import de.qabel.qabelbox.dagger.modules.FileBrowserModule;
+import de.qabel.qabelbox.box.dagger.modules.FileBrowserModule;
 import de.qabel.qabelbox.dagger.modules.MainActivityModule;
 import de.qabel.qabelbox.dagger.scopes.ActivityScope;
